cleaned up index.js and refactored othe functions into specific folders
