+++ frontend/src/Components/ProviderHomePage/ProviderHomePage.jsx
@@ -262,3 +262,4 @@ const ProviderHomePage = () => {
 };
 
 export default ProviderHomePage;
+
